+++ dorado/read_pipeline/ModBaseCallerNode.h
@@ -27,12 +27,12 @@ private:
     void output_worker_thread();
 
     // Worker threads, scales and chunks reads for callers and enqueues them
-    void runner_worker_thread(int runner_id);
+    void runner_worker_thread(size_t runner_id);
     // Worker thread per caller, performs the GPU calls to the remora models
-    void caller_worker_thread(int caller_id);
+    void caller_worker_thread(size_t caller_id);
 
     // Called by caller_worker_thread, calls the model and enqueues the results
-    void call_current_batch(int caller_id);
+    void call_current_batch(size_t caller_id);
 
     ReadSink& m_sink;
     size_t m_batch_size;
